pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "scientist"
version = "0.1.0"
description = "Safe refactoring through controlled experiments — a Python port of GitHub's Scientist"
requires-python = ">=3.11"
license = "MIT"
dependencies = []

[project.optional-dependencies]
otel = ["opentelemetry-api>=1.20.0"]
structlog = ["structlog>=24.0.0"]
all = ["opentelemetry-api>=1.20.0", "structlog>=24.0.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "structlog>=24.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/scientist"]

[tool.hatch.envs.default]
features = ["dev"]

[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"

[tool.pytest.ini_options]
asyncio_mode = "auto"
